+++ tests/AbstractTestCase.php
@@ -2,13 +2,49 @@
 
 namespace AvtoDev\Sentry\Tests;
 
-use Illuminate\Foundation\Testing\TestCase as BaseTestCase;
+use AvtoDev\Sentry\ServiceProvider;
+use Illuminate\Contracts\Console\Kernel;
+use Sentry\Laravel\ServiceProvider as SentryLaravelServiceProvider;
+use AvtoDev\AppVersion\ServiceProvider as AppVersionServiceProvider;
 
-/**
- * Class AbstractTestCase.
- */
-abstract class AbstractTestCase extends BaseTestCase
+abstract class AbstractTestCase extends \Illuminate\Foundation\Testing\TestCase
 {
-    use Traits\CreatesApplicationTrait,
-        Traits\AdditionalAssertsTrait;
+    /**
+     * Returns array of default service providers classes.
+     *
+     * @return string[]
+     */
+    public static function getDefaultServiceProviders(): array
+    {
+        return [
+            SentryLaravelServiceProvider::class,
+            AppVersionServiceProvider::class,
+            ServiceProvider::class,
+        ];
+    }
+
+    /**
+     * Creates the application.
+     *
+     * @param string ...$service_providers
+     *
+     * @return \Illuminate\Foundation\Application
+     */
+    public function createApplication()
+    {
+        /** @var \Illuminate\Foundation\Application $app */
+        $app = require __DIR__ . '/../vendor/laravel/laravel/bootstrap/app.php';
+
+        $app->make(Kernel::class)->bootstrap();
+
+        $service_providers = \func_num_args() >= 1
+            ? \func_get_args()
+            : self::getDefaultServiceProviders();
+
+        foreach ($service_providers as $service_provider) {
+            $app->register($service_provider);
+        }
+
+        return $app;
+    }
 }